Version 0.22
-Fixed Team Movement
-Fixed various Match Simulation bugs
-Added skip match functionality (could be better)
-Fixed bug where matches from other teams could be viewed. Eventually
this should be a feature, but for now is unnecessary and didn’t work
properly.
-Created functionality to expand athlete information on Team Panel.
Currently a trade button pops up that doesn’t work fully yet. Later can
include statistics and such.
-Trade athlete function is in place but currently nullifies an athlete
value, which leads to errors.
-Created team ranking function that’s called at the end of each week.
-Created bones of postseason system.
-Added season records to the end of team names

Design Improvements:
-Outlined an idea for a pretty slick sport. Currently undergoing paper
prototyping and testing. Won’t bother implementing until I’m certain it
will be positive. For now, I’d like to focus on the simulation and
management aspects since I know for a fact those will remain in the
game.