Version 0.23
-Implemented and reorganized postseason structure.
-All games for the season, including those in the postseason are
included in one singular jagged array. This allows me to use the same
AdvanceWeek and ConcludeWeek function throughout the postseason,
whereas earlier I had to create separate functions for the postseason.
This has greatly reduced the amount of repeated code and I was able to
remove 100 or so lines of shitty code.
-Restructured AdvanceWeek to only call TeamMovement for teams that are
in matches for the current week.
-The week county technically starts at -1. The week is now advanced
immediately after the player pressed the advance week button, instead
of at the conclusion of the week. This just made more sense to me and
now it’s a bit easier to understand. Still a little confusing because
Week 1 (for the player) in the code is still technically week 0. I’ll
probably leave it like that because outside of coding it doesn’t really
make sense to start at 0 for the first week.
-Fixed various bugs.
-Added plenty of bugs to fix later.