Version 0.24
-Removed placeholder simulation match mechanic.
-Created new match structure with several key features:
1) Each athlete on the field (currently 5 per team) will start with the
ball once per match based on an alternating order list.
2) Each Play consists of one Opportunity for a player. Opportunities
available to the athlete are based on athlete, teammate, and opponent
field position, whether they have the ball, and eventually their
stats/abilities. Currently these opportunities are: shoot, heave,
drive, dunk, pass, through, defense, protect, and steal. Each is stored
as an Opportunity object with values that determine the likelihood of
success. Steals steal an opponent’s turn in the same field position if
it’s successful.
3) When a player’s athlete begins a Play their name and stats are
highlighted on the TeamPanel and the available opportunities are
presented to the player as UI that can be selected. The opponent’s
choice is determined at random for now.
4) The previous match conclusion was left unchanged, highest score wins
and ties go to the home team for now.

-Added some more Getters and Setters including one to retrieve an
athlete’s team.
-Removed old code