Version 0.25
-Minor tweaks to match UI elements
-Match positions are now randomized between their defensive goal and
midfield. Later these positions should be assigned by the manager.
-Added third party name generator to get familiar with incorporating
unfamiliar code and because it’s super easy to implement and pretty
nifty. Currently this is only being used for the draft classes but I’ll
likely implement some form of it to be used universally.
-Cleaned up the Inspector with headers
-Created a draft function and UI elements, including a draft order
determined by season standings and a randomly generated draft class of
18.
-Created an AthleteUI prefab and script that is being used in the draft
class (will likely universalize). This script contains public
references to the name and stat texts of a particular athlete and has
mouse interaction to allow highlighting and clicking on the object.
-Created more Getters/Setters for TeamController to increase my
codifying abilities.
-Implemented step by step draft system that automates selections for AI
and then stops to wait for the player to click on and confirm their
selection before continuing the draft.
-Commented out unused code that will likely be used in the future.
-Created a debugger function to quickly test offseason operations.
-Noted down future ideas for quality gameplay involving royal decrees
and league voting.