Version 0.3
-UI is now much more space-efficient. Still not finalized but it’s
cleaner.
-New font for everything (which looks dope). Separate font is used for
match texts because the letters are very close together and allows for
more words. Temp font.
-Created AthleteUI objects and various functionality.
-Added ability to rename athletes.
-Added indicators to the match UI displaying which athletes are
selected and/or have the ball.
-Gave descriptions to each of the opportunity buttons available to the
player. Considering changing these to quotes from the athlete dependent
on their likelihood to want to do that action.
-Randomized the order of schedule creation. Still slightly uneven.
-Added functionality for flexible roster size both in TeamController
classes and UI elements.
-Fixed how goals, steals, and defense works. May change.
-Fixed athlete info display.
-Created a preseason period connecting the aftermath of the draft with
the start of the next season.
-Created functionality to play through multiple (potentially infinite)
seasons.
-Added week indicator text to top right corner that will likely almost
always be displayed.
-Replaced the select team button with a universal advance button.

-Created messages from the owner that appear before the start of each
season outlining expectations and general plot dialogue. After each
week there is a chance to get a message from the owner congratulating,
scolding, or offering a bounty for winning the next game. This should
add for a bit more flavor to the game as well as providing additional
choices, lore, goals, and rewards.

-Added flexible match active roster size functionality and lowered
starting number to 3.
-Added (untested) functionality to select their athletes before
starting a match.
-Began creation of FieldPosition objects and general field state.
-Began migrating match functionality into matchup class. Game stops as
soon as a match starts.
-Various code clean up and refactoring.

Known Bugs:
-Athlete renaming function has edge case.
-Occassionally the match panel is halfway off the screen for seemingly
no reason.