Version 0.31
-Began utilizing Trello for production management and prioritization.
-Finished migrating match functionality entirely to match script, which
is now about the same length as the GameController script.
-Implemented and player athlete lineup selection before matches.
Currently managers can be added to active rosters and used during
matches. Eventually this is something that can be changed via
postseason committee votes.
-Reworked field position to hold the global value, with certain parts
of code finding their relative field position.
-Reworked how matches work so that they flow from play to play
depending on the result of each opportunity attempt.
-Re-implemented brief pause before each opportunity attempt.
-Added halftime functionality that resets athlete positions.
-Added halftime and final generated match texts.
-Created functionality to allow stats to impact athlete’s percent
chance at success for each opportunity. This works by having a base
percentage and a max percentage, with the max achieved if the athlete
has a value of 10 for each stat that the opportunity uses.
-Each opportunity button displayed to the player now includes a small
set of quotes from the player that later will be more nuanced depending
on how likely they would be to want to attempt that particular
opportunity. This is displayed when the user hovers over the
opportunity button.

-Fixed ball indicator placement.