Version 0.32
-Reworked county/city definitions and functionality. Each city will
always have one team with this structure.
-Created an indicator during matches displaying how many more
possession there will be before halftime.
-Created stats for turnovers, boards, goals, steals, etc..
-Created board functionality after shots. All athletes in that goal
zone have an equal opportunity to snag the board.
-Created notifications displaying what happened on-field that appear
above each athlete’s head. Displays for goals, misses, passes, boards,
and turnovers.
-Commented out the old athlete info panel functionality. Created a new
athlete detail panel that looks cleaner and is directly on the team
panel instead of popping up off of it. Displays when hovering over an
athlete’s UI.
-Created different cities and teams for every county.
-Each county now has a county color that acts as the border color of
each team. This way teams have two colors and share a common border
color with other teams in their county.
-Implemented functionality to select non-capital teams in counties.
-Reworked the functionality of mouse interactions with counties and
cities to be more user-friendly and understandable.
-Created a back to world view button when the camera is focused on a
county.
-Team selection is now done by clicking on the city and then confirming
instead of clicking on the city and then clicking on another button and
then confirming.
-Created county panels that display when hovering over a county that
display the county’s capital and team. This also acts as the border for
the team panel.
-Changed matchup arrays to matchup lists to make it more flexible and
account for variable number of matches per week.
-Created leagues for each county that currently generate their
schedules the same way that the primary league does. This leads to an
issue if the county has an even number of teams (since the capital team
doesn’t participate in county leagues and the schedule generator
doesn’t work with odd numbers).
-Reworked TeamMovement functionality to get all county teams moving and
matches working properly.
-Created a Race class for Chookens and Lagunas that has a separate
image and jersey image displayed on athlete field UI and the athlete
detail panel. Field UI Jersey colors change based on team color.
-Created Chooken and Laguna placeholder art assets.